% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMLR.weight.R
\name{CIMLR.weight}
\alias{CIMLR.weight}
\title{CIMLR.weight}
\usage{
CIMLR.weight(X, c, no.dim = NA, k = 10, cores.ratio = 0, weight)
}
\arguments{
\item{X}{a list of multi-omic data each of which is an (n x m) data matrix of measurements of cancer patients}

\item{c}{number of clusters to be estimated over X}

\item{no.dim}{number of dimensions}

\item{k}{tuning parameter}

\item{cores.ratio}{ratio of the number of cores to be used when computing the multi-kernel}

\item{weight}{a list of weight for multi-omic data, each of which is an m dimensional vector}
}
\value{
clusters the patients based on CIMLR.weight and their similarities

list of 8 elements describing the clusters obtained by CIMLR, of which y are the resulting clusters:
     y = results of k-means clusterings,
     y_spectral=results of spectral clustering,
     S = similarities computed by CIMLR,
     F = results from tsne(S),
     ydata = data referring the the results by k-means,
     alphaK = clustering coefficients,
     execution.time = execution time of the present run,
     converge = iterative convergence values by T-SNE,
     LF = parameters of the clustering
}
\description{
perform the CIMLR.weight clustering algorithm
}
\examples{
CIMLR.weight(X =exampledata$PRCCReduced_X, c = 3,
             cores.ratio = 0,weight=exampledata$weightforfeatures)

}
