% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata.R
\docType{data}
\name{exampledata}
\alias{exampledata}
\title{test dataset for wMLR}
\format{
multi-omic data of cancer patients
}
\usage{
data(exampledata)
}
\value{
list of 2 element:
	PRCCReduced_X  = input dataset as a list of 4 (reduced) multi-omic data each of which is an (n x m) measurements of cancer patients
   weightforfeatures=input dataset as a list of 4 (reduced) weight for multi-omic data each of which is an m measurements of features
}
\description{
example dataset to test wMLR. This is a reduced version of the dataset papillary renal cell carcinoma (PRCC)
             in weight-boosted Multikernel LeaRning (wMLR) paper.
}
